+++ src/Regressors/GradientBoost.php
@@ -128,7 +128,7 @@ class GradientBoost implements Estimator, Learner, Iterative, RanksFeatures, Ver
      * The number of evaluations without improvement in the validation score to wait before considering an
      * early stop.
      *
-     * @var positive-int
+     * @var int<0,max>
      */
     protected int $window;
 
@@ -236,7 +236,7 @@ class GradientBoost implements Estimator, Learner, Iterative, RanksFeatures, Ver
                 . " greater than 0, $evalInterval given.");
         }
 
-        if ($window < 1) {
+        if ($window < 0) {
             throw new InvalidArgumentException('Window must be'
                 . " greater than 0, $window given.");
         }
@@ -403,7 +403,7 @@ class GradientBoost implements Estimator, Learner, Iterative, RanksFeatures, Ver
         $this->mu = $mu;
 
         $bestScore = $minScore;
-        $bestEpoch = $numWorseEpochs = 0;
+        $bestEpoch = $numWorseEvals = 0;
         $score = null;
         $prevLoss = INF;
 
@@ -445,6 +445,11 @@ class GradientBoost implements Estimator, Learner, Iterative, RanksFeatures, Ver
 
             if ($evalThisStep) {
                 if ($score >= $maxScore) {
+                    if ($this->logger) {
+                        $this->logger->info('Early stopping, maximum '
+                            . "{$this->metric} score reached");
+                    }
+
                     break;
                 }
 
@@ -452,17 +457,27 @@ class GradientBoost implements Estimator, Learner, Iterative, RanksFeatures, Ver
                     $bestScore = $score;
                     $bestEpoch = $epoch;
 
-                    $numWorseEpochs = 0;
+                    $numWorseEvals = 0;
                 } else {
-                    ++$numWorseEpochs;
+                    ++$numWorseEvals;
                 }
 
-                if ($numWorseEpochs >= $this->window) {
+                if ($this->window and $numWorseEvals >= $this->window) {
+                    if ($this->logger) {
+                        $this->logger->info('Early stopping, no improvement in '
+                            . "the last {$this->window} evaluations");
+                    }
+
                     break;
                 }
             }
 
             if ($lossChange < $this->minChange) {
+                if ($this->logger) {
+                    $this->logger->info('Early stopping, loss change below '
+                        . "minimum of {$this->minChange}");
+                }
+
                 break;
             }