+++ tests/Classifiers/MultilayerPerceptronTest.php
@@ -125,6 +125,38 @@ class MultilayerPerceptronTest extends TestCase
     }
 
     #[Test]
+    public function windowDisabled() : void
+    {
+        srand(self::RANDOM_SEED);
+
+        $estimator = new MultilayerPerceptron(
+            hiddenLayers: [
+                new Dense(16),
+                new Activation(new LeakyReLU(0.1)),
+                new Dense(8),
+                new Activation(new SoftPlus()),
+            ],
+            batchSize: 16,
+            optimizer: new Adam(new Constant(0.01)),
+            epochs: 10,
+            minChange: 1e-12,
+            evalInterval: 1,
+            window: 0,
+            holdOut: 0.1,
+            costFn: new MulticlassCrossEntropy(),
+            metric: new FBeta()
+        );
+
+        $estimator->setLogger(new BlackHole());
+
+        $training = $this->generator->generate(self::TEST_SIZE);
+
+        $estimator->train($training);
+
+        $this->assertTrue($estimator->trained());
+    }
+
+    #[Test]
     public function progressContract() : void
     {
         $this->assertSame([], iterator_to_array($this->estimator->progress(), false));